refactor(cockpit): extract composition module
